+++ w13_kontrolltoo2/jooned/src/main/java/org/kt/JoonLiides.java
@@ -1,6 +1,6 @@
 package org.kt;
 
 public interface JoonLiides {
-    public double getY(double x);
+    public int getY(int x);
 }
